[cr132] Satisfy `borrowed-range` use with `base::span`
In M132 `base::span` underwent a series of improvements to make type
deduction better. One of them is to satisfy the requirement for borrowed
range, to avoid the use of dangling spans.

This change corrects a few cases where the `borrowed_range` check is
failing, but the span is being used in the same expression. This
requires to change the span type to const, so `base::as_span_bytes` is
used in a few places.