[Android] AccessibilitySettings moved to components
Chromium change:
https://chromium.googlesource.com/chromium/src/+/1598764f95396242e0ddfc47e43e5133c2bd8da3

Move AccessibilitySettings to the accessibility component

This CL moves the AccessibilitySettings class to the accessibility
component: //components/browser_ui/accessibility

The class has previously been tied to //chrome/browser, even though
there was a delegate to provide embedder-specific implementations. We
have moved all deps to the delegate, and now AccessibilitySettings
depends only on //content_public/ and //components/, and we can move
this to be alongside the rest of the accessibility browser_ui files.

AX-Relnotes: N/A
Bug: 379128588